+++ pds/src/main/java/dominio/info/usuario/infoEstadisticasProfesor.java
@@ -7,7 +7,7 @@ public class infoEstadisticasProfesor extends infoEstadisticas{
     
     private int cursosPublicados;
 
-    public infoEstadisticasProfesor (String texto, int puntuacion, int cursosCompletados, int bloquesCompletados, Duration minutosUso, int diasUso, int maxRacha, int cursosPublicados){
+    public infoEstadisticasProfesor (String texto, double puntuacion, int cursosCompletados, int bloquesCompletados, Duration minutosUso, int diasUso, int maxRacha, int cursosPublicados){
         super(texto, puntuacion, cursosCompletados, bloquesCompletados, minutosUso, diasUso, maxRacha);
         
         this.cursosPublicados = cursosPublicados;
